+++ {{ cookiecutter.directory_name }}/tests/test.py
@@ -1,8 +1,0 @@
-import unittest
-
-
-class Test(unittest.TestCase):
-
-    def test_true_is_true(self):
-        self.assertEqual(True, True)
-        # Your test case here